requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
Babel==2.9.0
cfgv==3.2.0
distlib==0.3.1
Django==3.1.7
django-axes==5.13.1
django-extensions==3.1.1
django-ipware==3.0.2
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-reversion==3.0.9
django-smuggler==1.0.2
django-unused-media==0.2.2
filelock==3.0.12
identify==2.2.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
nodeenv==1.5.0
phonenumbers==8.12.20
Pillow==8.1.2
pre-commit==2.11.1
psycopg2==2.8.6
pyflakes==2.3.1
pygraphviz==1.7
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
reportlab==3.5.66
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.3
wrapt==1.12.1
